soft delete aggiunta per i prodotti e modifiche varie nei template e nell'interfaccia admin

+++ src/Foundation/FImmagine.php
@@ -45,8 +45,10 @@ class FImmagine extends EntityRepository {
         $em = getEntityManager();
         $found_prodotto = $em->find(EProdotto::class, $productId);
         $found_images = self::getAllObjectImages($found_prodotto);
-        foreach($found_images as $image){
-            $em->remove($image);
+        if(!$found_prodotto->isDeleted()){
+            foreach($found_images as $image){
+                $em->remove($image);
+            }
         }
         $em->flush();
     }
